+++ src/main/java/com/shoppingcartservice/sachin/Reposistories/Cart/CartRepo.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartRepo extends JpaRepository<Cart,Long> {
-    Cart findCartByCartId(Long cartId);
+public interface CartRepo extends JpaRepository<Cart,Integer> {
+    Cart findCartByCartId(Integer cartId);
     Cart findCartByUserProfile(UserProfile userProfile);
 }
